LayoutTests/java/lc3/Constructors/construct-001-expected.txt

CONSOLE MESSAGE: line 37: ReferenceError: Can't find variable: Packages
LiveConnect 3.0 JavaScript to Java Data Conversion

FAIL successfullyParsed should be true. Threw exception ReferenceError: Can't find variable: successfullyParsed

TEST COMPLETE
 
